+++ exchange/errors.py
@@ -1,6 +1,3 @@
 class ExchangeError(Exception):
     pass
 
-
-class ExchangeConfigError(ExchangeError):
-    pass