+++ exchange/urls.py
@@ -3,7 +3,10 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    # TODO
-    # path('', views.convert, name='convert'),
-    # path('', views.rates, name='rates'),
+    path('rates/<str:base_ccy>',
+         views.rates,
+         name='rates'),
+    path('convert/<str:base_ccy>/<str:to_ccy>/<str:amount>',
+         views.convert,
+         name='convert'),
 ]